requirements.txt
# ===================================
# Diaco MES - Production Management
# شرکت توسعه هوشمند فرش ایرانیان
# ===================================

# Django Core (LTS - سازگار با MariaDB 10.4)
Django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

# Database
mysqlclient==2.2.7

# Jalali Calendar
django-jalali==7.1.0
jdatetime==5.0.0

# Utilities
django-filter==24.3
django-cors-headers==4.6.0
Pillow==11.1.0
python-decouple==3.8

# Development
django-debug-toolbar==4.4.6
